apps/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]<0.116.0",
    "litellm[proxy]>=1.74.8",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
]
